+++ app/modules/periodic_report/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.6.0'
+__version__ = '0.7.0'
 
 from .routes import periodic_report_bp
 from .service import PeriodicReportService